+++ app/Http/Controllers/Educacion/InstitucionCertificadoraController.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace App\Http\Controllers\Educacion;
+
+use App\Models\InstitucionCertificadora;
+use Illuminate\Http\Request;
+
+class InstitucionCertificadoraController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\InstitucionCertificadora  $institucionCertificadora
+     * @return \Illuminate\Http\Response
+     */
+    public function show(InstitucionCertificadora $institucionCertificadora)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Models\InstitucionCertificadora  $institucionCertificadora
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, InstitucionCertificadora $institucionCertificadora)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Models\InstitucionCertificadora  $institucionCertificadora
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(InstitucionCertificadora $institucionCertificadora)
+    {
+        //
+    }
+}
