+++ app/Http/Controllers/Educacion/TipoCertificacionController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Perfil;
+namespace App\Http\Controllers\Educacion;
 
 use App\Models\Perfil\TipoCertificacion;
 use Illuminate\Http\Request;
