+++ app/Models/Educacion/InstitucionCertificadora.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models\Educacion;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class InstitucionCertificadora extends Model
+{
+    use HasFactory;
+}
